Strategy now check if a player is ban by the double touch detector, before kicking
