+++ ai/STA/Strategy/offense.py
@@ -5,7 +5,8 @@ from Util.constant import KEEPOUT_DISTANCE_FROM_GOAL
 from Util.geometry import Area, normalize
 from Util.role import Role
 
-from ai.Algorithm.evaluation_module import closest_player_to_point, closest_players_to_point
+from ai.Algorithm.evaluation_module import closest_player_to_point, closest_players_to_point, \
+    closest_players_to_point_except
 from ai.STA.Strategy.strategy import Strategy
 from ai.STA.Tactic.go_kick import GoKick
 from ai.STA.Tactic.goalkeeper import GoalKeeper
@@ -66,11 +67,14 @@ class Offense(Strategy):
                 Role.SECOND_DEFENCE]
 
     def is_closest_not_goalkeeper(self, player):
-        closest_players = closest_players_to_point(self.game_state.ball_position, our_team=True)
-        if player == closest_players[0].player:
-            return True
-        return closest_players[0].player == self.game_state.get_player_by_role(Role.GOALKEEPER) \
-               and player == closest_players[1].player
+        ban_players = self.game_state.double_touch_checker.ban_players
+        if player in ban_players:
+            return False
+
+        closests = closest_players_to_point_except(self.game_state.ball.position,
+                                                   except_roles=[Role.GOALKEEPER],
+                                                   except_players=ban_players)
+        return len(closests) > 0 and closests[0].player == player
 
     def is_not_closest(self, player):
         return not self.is_closest_not_goalkeeper(player)
